ENG-11012: Online Tracing System.
Introducing an online tracing system that can be turned on and off at
runtime without significantly impacting the workload. Trace events will
be cached in a ring buffer in memory unless instructed to be written to
file. Since the ring buffer is bounded, old trace events will be removed
if the buffer is filled up. This allows tracing to be turned on without
worrying about running out of heap. The ring buffer size can be tweaked
using the Java system property "VOLTTRACE_QUEUE_SIZE". By default, it is
set to 4096.

The tracing system can be turned on/off for each part of VoltDB's
subsystems individually. Trace events are currently available for the
following subsystems.

- CI: Client Interface.
- MPI: Multi-partition Initiator.
- MPSITE: Multi-partition Site.
- SPI: Single-partition Initiator.
- SPSITE: Single-partition Site.
- EE: Execution Engine.
- DRPRODUCER: DR Producer.
- DRCONSUMER: DR Consumer.

The trace points available now are sufficient to diagnose most
transaction related problems such as slow query/transaction, overhead
incurred by command logging (sync or async), or spikes in transaction
latency, etc. More trace points can be added to other parts of the
system.

The output file format is JSON. The JSON content conforms to the Chrome
Trace Event format, so it can easily be loaded into Chrome's tracing
tool for visualization. It is also easy to parse and analyze in
programs.

I added a sysproc to manage the tracing system at runtime. It has the
following parameters.

- @Trace on: Turn on the tracing system if it's not turned on already.
- @Trace off: Turn the tracing system off if it is on.
- @Trace dump: Write all trace events in the ring buffer to file and
  return the file path.
- @Trace enable category: Enable tracing for the given category.
- @Trace disable category: Disable tracing for the given category.

SQLCmd currently recognizes the @Trace sysproc, but not VMC. It could
also be added to voltadmin for ease of use.

Based on my preliminary benchmark result, turning the tracing system on
for all existing categories doesn't seem to add noticeable overhead to
throughput but adds several milliseconds to the latency. If tracing is
turned off, it adds almost zero overhead to the system.