Add a test for ENG-12076.
Also, fix up the test for ENG-8280.  The column resolution code for
AbstractJoinPlanNode does some resolution for a subquery, and then tries
to resolve taking into account inline plan nodes.  The subquery join
resolution was broken, and set the indexes and column differentiators
for the inner join branch to be the same as for the outer join branch.
This nonsense is fixed here.